Testes unitários de sucesso e falha.
