+++ fontes/interfaces/erros/erro-analisador-semantico.ts
@@ -3,7 +3,6 @@ import { SimboloInterface } from "../simbolo-interface";
 export interface ErroAnalisadorSemantico {
     simbolo?: SimboloInterface;
     mensagem?: string;
-    erroInterno?: any;
     linha?: number;
     hashArquivo?: number;
 }
