+++ fontes/interfaces/retornos/retorno-analisador-semantico.ts
@@ -2,5 +2,4 @@ import { ErroAnalisadorSemantico } from "../erros";
 
 export interface RetornoAnalisadorSemantico {
     erros: ErroAnalisadorSemantico[];
-    resultado: string[];
 }
